ai-report/requirements.txt
psycopg[binary]>=3.1
jinja2>=3.1
python-dotenv>=1.0
openai>=1.30